move matrices to their own header
